Fixing compiler warning "comparison between ‘enum <anonymous>’ and ‘enum <anonymous>’" when compiling with gcc >= 4.5

+++ bertos/drv/i2c.h
@@ -435,7 +435,7 @@ INLINE int i2c_error(I2c *i2c)
  *            and similar if you want to activate the bitbang driver.
  * \param clock Peripheral clock
  */
-#define i2c_init_3(i2c, dev, clock)   ((((dev) >= I2C_BITBANG0) | ((dev) == I2C_BITBANG_OLD)) ? \
+#define i2c_init_3(i2c, dev, clock)   ((((dev) >= (int)I2C_BITBANG0) | ((dev) == (int)I2C_BITBANG_OLD)) ? \
 										i2c_hw_bitbangInit((i2c), (dev)) : i2c_hw_init((i2c), (dev), (clock)))
 /**@}*/
 /**\}*/ // i2c_api
